Finished refactoring - broke out chat backend and gui.
